moved all HTTP get functions into modules
