+++ src/main.rs
@@ -2,16 +2,14 @@ mod cmd;
 mod core;
 mod http;
 mod model;
+mod run;
 mod util;
 
 use crate::cmd::cli::{CliOptions, OutputOption};
 use crate::cmd::command::CliCommand;
 use crate::core::config::ExtraHopConfig;
-use crate::http::activity_map::get_activity_maps;
-use crate::http::api_key::get_api_keys;
-use crate::http::audit_log::get_audit_logs;
+use crate::http::action::*;
 use crate::http::client::{build_clients, ExtraHopAppliance, ExtraHopClient};
-use crate::http::common::reqwest_get;
 use crate::http::firmware::{get_firmware_next, get_firmware_previous, upload_firmware};
 use crate::http::getter::{appliance_getters, GetterType};
 
@@ -37,342 +35,17 @@ use model::network::Networks;
 use model::network_locality::NetworkLocalities;
 use model::node::Nodes;
 use model::packet_capture::PacketCaptures;
-use model::packet_search::PacketSearch;
-use model::running_config::RunningConfig;
 use model::software::Softwares;
 use model::tag::Tags;
 use model::threat_collection::ThreatCollections;
 use model::trigger::Triggers;
 use model::vlan::Vlans;
 
-use anyhow::{Context, Result};
+use anyhow::Result;
 use chrono::Local;
-use futures_util::StreamExt;
-use reqwest::StatusCode;
 use std::collections::HashMap;
-use std::fs::File;
-use std::io::Write;
 use std::process::exit;
 use tabled::{Alignment, Disable, Full, MaxWidth, MinWidth, Modify, Rotate, Rows, Table};
-
-async fn get_alerts(client: &ExtraHopClient) -> Result<Alerts> {
-    let response = reqwest_get(client, "alerts").await?;
-    let alerts = Alerts {
-        alerts: serde_json::from_str(&response.text().await?)?,
-    };
-    Ok(alerts)
-}
-
-async fn get_appliances(client: &ExtraHopClient) -> Result<Appliances> {
-    let response = reqwest_get(client, "appliances").await?;
-    let appliances = Appliances {
-        appliances: serde_json::from_str(&response.text().await?)?,
-    };
-    Ok(appliances)
-}
-
-async fn get_bundles(client: &ExtraHopClient) -> Result<Bundles> {
-    let response = reqwest_get(client, "bundles").await?;
-    let bundles = Bundles {
-        bundles: serde_json::from_str(&response.text().await?)?,
-    };
-    Ok(bundles)
-}
-
-async fn get_dashboards(client: &ExtraHopClient) -> Result<Dashboards> {
-    let response = reqwest_get(client, "dashboards").await?;
-    let dashboards = Dashboards {
-        dashboards: serde_json::from_str(&response.text().await?)?,
-    };
-    Ok(dashboards)
-}
-
-async fn get_detections(client: &ExtraHopClient) -> Result<Detections> {
-    let response = reqwest_get(client, "detections").await?;
-    let detections = Detections {
-        detections: serde_json::from_str(&response.text().await?)?,
-    };
-    Ok(detections)
-}
-
-async fn get_devices(client: &ExtraHopClient) -> Result<Devices> {
-    let response = reqwest_get(client, "devices").await?;
-    let devices = Devices {
-        devices: serde_json::from_str(&response.text().await?)?,
-    };
-    Ok(devices)
-}
-
-async fn get_device_groups(client: &ExtraHopClient) -> Result<DeviceGroups> {
-    let response = reqwest_get(client, "devicegroups").await?;
-    let device_groups = DeviceGroups {
-        device_groups: serde_json::from_str(&response.text().await?)?,
-    };
-    Ok(device_groups)
-}
-
-async fn get_customizations(client: &ExtraHopClient) -> Result<Customizations> {
-    let response = reqwest_get(client, "customizations").await?;
-    let customizations = Customizations {
-        customizations: serde_json::from_str(&response.text().await?)?,
-    };
-    Ok(customizations)
-}
-
-async fn create_customization(client: &ExtraHopClient) -> Result<()> {
-    let name = format!("{}-{}", client.hostname, client.timestamp);
-    let body = serde_json::json!({ "name": name.to_string() });
-
-    println!("=> adding customization: {}", name);
-    let url = format!("{}/customizations", client.base_url);
-    let response = client.reqwest_client.post(url).json(&body).send().await?;
-
-    match response.status() {
-        StatusCode::CREATED => {
-            println!("=> new customization added: {}", name);
-            let customizations = get_customizations(client).await?;
-            for c in customizations.customizations.iter() {
-                if c.name.starts_with(&name) {
-                    save_customization(client, &c.id).await?;
-                }
-            }
-            Ok(())
-        }
-        _ => {
-            eprintln!("=> unable to add customization: {}", name);
-            eprintln!("{:#?}", response.error_for_status());
-            exit(1)
-        }
-    }
-}
-
-async fn save_customization(client: &ExtraHopClient, id: &u64) -> Result<()> {
-    let name = format!("{}-{}", client.hostname, client.timestamp);
-    let url = format!("{}/customizations/{}/download", client.base_url, id);
-    let response = client.reqwest_client.post(url).send().await?;
-
-    match response.status() {
-        StatusCode::OK => {
-            println!("=> downloading customization: {}", name);
-            let bytes = response.bytes().await?;
-            let filename = format!("{}-{}.zip", client.hostname, client.timestamp);
-            let mut wf = File::create(&filename)?;
-            wf.write_all(&bytes)
-                .expect("=> error writing customization to file");
-            Ok(())
-        }
-        _ => {
-            eprintln!("=> unable to save customization: {}", name);
-            eprintln!("=> {:#?}", response.error_for_status());
-            exit(1)
-        }
-    }
-}
-
-async fn get_custom_devices(client: &ExtraHopClient) -> Result<CustomDevices> {
-    let response = reqwest_get(client, "customdevices").await?;
-    let custom_devices = CustomDevices {
-        custom_devices: serde_json::from_str(&response.text().await?)?,
-    };
-    Ok(custom_devices)
-}
-
-async fn get_email_groups(client: &ExtraHopClient) -> Result<EmailGroups> {
-    let response = reqwest_get(client, "emailgroups").await?;
-    let email_groups = EmailGroups {
-        email_groups: serde_json::from_str(&response.text().await?)?,
-    };
-    Ok(email_groups)
-}
-
-async fn get_exclusion_intervals(client: &ExtraHopClient) -> Result<ExclusionIntervals> {
-    let response = reqwest_get(client, "exclusionintervals").await?;
-    let exclusion_intervals = ExclusionIntervals {
-        exclusion_intervals: serde_json::from_str(&response.text().await?)?,
-    };
-    Ok(exclusion_intervals)
-}
-
-async fn get_extrahop(client: &ExtraHopClient) -> Result<ExtraHop> {
-    let response = reqwest_get(client, "extrahop").await?;
-    let extrahop: ExtraHop = serde_json::from_str(&response.text().await?)?;
-    Ok(extrahop)
-}
-
-async fn get_running_config(client: &ExtraHopClient) -> Result<()> {
-    let response = reqwest_get(client, "runningconfig").await?;
-
-    match response.status() {
-        StatusCode::OK => {
-            let json_data: serde_json::Value = serde_json::from_str(&response.text().await?)?;
-
-            let config = RunningConfig { json: json_data };
-
-            let filename = format!("{}-{}.json", client.hostname, client.timestamp);
-            let wf = serde_json::to_writer(&File::create(&filename)?, &config.json);
-            match wf {
-                Ok(_) => println!("=> wrote file: {}", &filename),
-                Err(_) => exit(1),
-            };
-            Ok(())
-        }
-        _ => {
-            eprintln!("=> unable to get running config");
-            eprintln!("=> {:#?}", response.error_for_status());
-            exit(1)
-        }
-    }
-}
-
-async fn get_identitiy_providers(client: &ExtraHopClient) -> Result<IdentitiyProviders> {
-    let response = reqwest_get(client, "/auth/identityproviders").await?;
-    let identity_providers = IdentitiyProviders {
-        identity_providers: serde_json::from_str(&response.text().await?)?,
-    };
-    Ok(identity_providers)
-}
-
-async fn get_license(client: &ExtraHopClient) -> Result<License> {
-    let response = reqwest_get(client, "license").await?;
-    let license: License = serde_json::from_str(&response.text().await?)?;
-    Ok(license)
-}
-
-async fn get_networks(client: &ExtraHopClient) -> Result<Networks> {
-    let response = reqwest_get(client, "networks").await?;
-    let networks = Networks {
-        networks: serde_json::from_str(&response.text().await?)?,
-    };
-    Ok(networks)
-}
-
-async fn get_network_localities(client: &ExtraHopClient) -> Result<NetworkLocalities> {
-    let response = reqwest_get(client, "networklocalities").await?;
-    let network_localities = NetworkLocalities {
-        network_localities: serde_json::from_str(&response.text().await?)?,
-    };
-    Ok(network_localities)
-}
-
-async fn get_nodes(client: &ExtraHopClient) -> Result<Nodes> {
-    let response = reqwest_get(client, "nodes").await?;
-    let nodes = Nodes {
-        nodes: serde_json::from_str(&response.text().await?)?,
-    };
-    Ok(nodes)
-}
-
-async fn get_packet_captures(client: &ExtraHopClient) -> Result<PacketCaptures> {
-    let response = reqwest_get(client, "packetcaptures").await?;
-    let packet_captures = PacketCaptures {
-        packet_captures: serde_json::from_str(&response.text().await?)?,
-    };
-    Ok(packet_captures)
-}
-
-async fn packet_search(client: &ExtraHopClient, options: &PacketSearch) -> Result<()> {
-    let name = format!("{}-{}", client.hostname, client.timestamp);
-    let filename = format!("{}.{}", name, options.output);
-    let url = format!("{}/packets/search", client.base_url);
-
-    let params = (
-        ("always_return_body", options.always_return_body),
-        ("bpf", options.bpf.clone()),
-        ("from", options.from.clone()),
-        ("ip1", options.ip1.clone()),
-        ("ip2", options.ip2.clone()),
-        ("limit_bytes", options.limit_bytes.clone()),
-        (
-            "limit_search_duration",
-            options.limit_search_duration.clone(),
-        ),
-        ("output", options.output.clone()),
-        ("port1", options.port1.clone()),
-        ("port2", options.port2.clone()),
-        ("until", options.until.clone()),
-    );
-
-    let response = client
-        .reqwest_client
-        .post(&url)
-        .form(&params)
-        .send()
-        .await?;
-
-    match response.status() {
-        StatusCode::OK => {
-            println!("=> downloading packets to `{}`", &filename);
-            let mut file = File::create(&filename)
-                .with_context(|| format!("=> failed to create `{}`", &filename))?;
-
-            let mut stream = response.bytes_stream();
-
-            while let Some(item) = stream.next().await {
-                let chunk =
-                    item.with_context(|| format!("=> error while downloading `{}`", &filename))?;
-                file.write_all(&chunk)
-                    .with_context(|| format!("=> error while writing to `{}`", &filename))?;
-            }
-        }
-        StatusCode::NO_CONTENT => {
-            println!("=> no packets returned from query")
-        }
-        _ => {
-            eprintln!("=> unable to save packets to `{}`", &filename);
-            eprintln!("=> {:#?}", response.error_for_status());
-            exit(1)
-        }
-    }
-    Ok(())
-}
-
-async fn get_saml_sp(client: &ExtraHopClient) -> Result<SamlSps> {
-    let response = reqwest_get(client, "/auth/samlsp").await?;
-    let saml_sps = SamlSps {
-        saml_sps: serde_json::from_str(&response.text().await?)?,
-    };
-    Ok(saml_sps)
-}
-
-async fn get_software(client: &ExtraHopClient) -> Result<Softwares> {
-    let response = reqwest_get(client, "software").await?;
-    let software = Softwares {
-        softwares: serde_json::from_str(&response.text().await?)?,
-    };
-    Ok(software)
-}
-
-async fn get_tags(client: &ExtraHopClient) -> Result<Tags> {
-    let response = reqwest_get(client, "tags").await?;
-    let tags = Tags {
-        tags: serde_json::from_str(&response.text().await?)?,
-    };
-    Ok(tags)
-}
-
-async fn get_threat_collections(client: &ExtraHopClient) -> Result<ThreatCollections> {
-    let response = reqwest_get(client, "threatcollections").await?;
-    let threat_collections = ThreatCollections {
-        threat_collections: serde_json::from_str(&response.text().await?)?,
-    };
-    Ok(threat_collections)
-}
-
-async fn get_triggers(client: &ExtraHopClient) -> Result<Triggers> {
-    let response = reqwest_get(client, "triggers").await?;
-    let triggers = Triggers {
-        triggers: serde_json::from_str(&response.text().await?)?,
-    };
-    Ok(triggers)
-}
-
-async fn get_vlans(client: &ExtraHopClient) -> Result<Vlans> {
-    let response = reqwest_get(client, "vlans").await?;
-    let vlans = Vlans {
-        vlans: serde_json::from_str(&response.text().await?)?,
-    };
-    Ok(vlans)
-}
 
 #[tokio::main]
 async fn main() -> Result<()> {
